js/xpconnect/tests/chrome/test_bug1126911.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1126911
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1126911</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1126911 **/
  var sb = new Cu.Sandbox(null);
  sb.win = window;
  sb.loc = location;
  function checkThrows(expr) {
    try {
      Cu.evalInSandbox(expr, sb);
      ok(false, "Should have thrown: " + expr);
    } catch (e) {
      ok(/denied|insecure/.test(e), "should get security exception: " + e);
    }
  }
  checkThrows('win.top');
  checkThrows('loc.replace');

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1126911">Mozilla Bug 1126911</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
